requirements.txt
prettytable==3.10.0
mysql-connector-python==8.0.29